backend/requirements.txt
Flask>=3.0,<4
Flask-Cors>=4.0,<6
python-dotenv>=1.0,<2
psycopg[binary]>=3.2,<4
PyJWT>=2.9,<3
gunicorn>=22.0,<24
